+++ scripts/config.js
@@ -4925,7 +4925,7 @@ export function setConfig() {
             },
             'Radiant Weapon': {
                 'name': 'Radiant Weapon',
-                'version': '0.7.18'
+                'version': '0.9.24'
             },
             'Repeating Shot': {
                 'name': 'Repeating Shot',
@@ -5682,6 +5682,22 @@ export function setConfig() {
                 'mutation': {
                     'self': 'Investiture of Stone'
                 }
+            },
+            'Create Pact Weapon': {
+                'name': 'Create Pact Weapon',
+                'version': '0.9.24'
+            },
+            'Pact Boon: Pact of the Blade': {
+                'name': 'Pact Boon: Pact of the Blade',
+                'version': '0.9.24'
+            },
+            'Eldritch Invocations: Improved Pact Weapon': {
+                'name': 'Eldritch Invocations: Improved Pact Weapon',
+                'version': '0.9.24'
+            },
+            'Hex Warrior': {
+                'name': 'Hex Warrior',
+                'version': '0.9.24'
             }
         }
     });